linhht_done

+++ src/Application/Commands/Employees/DeleteEmployeeCommand.cs
@@ -21,6 +21,8 @@ namespace Application.Commands.Employees
 
         public async Task<Response<int>> Handle(DeleteEmployeeCommand command, CancellationToken cancellationToken)
         {
+            if (command.Id == 1)//ADMIN prevent delete
+                throw new ApiException($"Role ADMIN couldnt be deleted");
             var entity = await _repository.GetByIdAsync(command.Id) ?? throw new ApiException($"Employee Not Found (Id:{command.Id}).");
             await _repository.DeleteEmployeeAsync(entity);
             return new Response<int>(entity.Id);
